rename: 산책 관련 뷰 walk 폴더로 이동
